LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_window_open_download_block_downloads.tentative-expected.txt

CONSOLE MESSAGE: Not allowed to download due to sandboxing
CONSOLE MESSAGE: Not allowed to download due to sandboxing


PASS window.open(download, ) triggering download in sandbox is blocked.
PASS window.open(download, "_blank") triggering download in sandbox is blocked.
PASS window.open(download, "_blank", "noopener") triggering download in sandbox is blocked.
